requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==11.1.0
PyPDF2==3.0.1
